Unify reply operation persistence into single collection and shared adapter
Merge coach_reply_operations and coach_conversation_reply_operations
collections into one llm_reply_operations collection with a scope_type
discriminator. Extract shared MongoLlmReplyOperationRepository that
implements both CoachReplyOperationRepository and
CoachConversationReplyOperationRepository.
