+++ src/adapters/mongo/mod.rs
@@ -7,15 +7,14 @@ pub mod calendar_entry_views;
 pub mod calendar_planned_workouts;
 pub mod client;
 pub mod coach_conversation_messages;
-pub mod coach_conversation_reply_operations;
 pub mod coach_conversations;
-pub mod coach_reply_operations;
 pub mod completed_workouts;
 mod error;
 pub mod external_observations;
 pub mod external_sync_states;
 pub mod ftp_history;
 pub mod llm_context_cache;
+pub mod llm_reply_operations;
 pub mod login_state;
 pub mod pest_parser_poc_workouts;
 pub mod planned_completed_links;
